Moved test to the test folder. <organisation
Moved Entitytest to the test folder, however I did not move PlayerFishTest
because I could not figure out how to stop the application that is used to
test the class.
