+++ src/main/Entity.java
@@ -24,7 +24,7 @@ public class Entity {
 	}
 	
 	public boolean intersectsRightScreenEdge() {
-		return (this.sprite.getAabb().getX() + this.sprite.getAabb().getWidth()) >= Game.resX;
+		return (this.sprite.getAabb().getX() + this.sprite.getAabb().getWidth()) >= Game.getResX();
 	}
 	
 	public boolean intersectsUpperScreenEdge() {
@@ -32,7 +32,7 @@ public class Entity {
 	}
 	
 	public boolean intersectsUnderScreenEdge() {
-		return (this.sprite.getAabb().getY() + this.sprite.getAabb().getHeight()) >= Game.resY;
+		return (this.sprite.getAabb().getY() + this.sprite.getAabb().getHeight()) >= Game.getResY();
 	}
 	
 	public Sprite getSprite() {
