+++ src/main/PlayerFishTest.java
@@ -1,11 +1,12 @@
 package main;
 
 import static org.junit.Assert.*;
-
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
+
 import static org.mockito.Mockito.mock;
+import javafx.application.Platform;
 import javafx.scene.image.Image;
 
 public class PlayerFishTest {
@@ -27,7 +28,7 @@ public class PlayerFishTest {
 	
 	@After
 	public void tearDown() {
-		
+		Platform.exit();
 	}
 
 	@Test
